move type to shared folder
