+++ src/connect/shared/index.ts
@@ -1,0 +1,2 @@
+export * from './Stats';
+export * from './Locale';
